Consolidate Criterion benchmarks
